adding basic framework for meshlet culling
- added a task shader (just a pass through for now)
    - Don't expect it to be faster than compute, but it is more
      convenient and I would like to see the perf
- models now calculate and export meshlet bounds + culling info
- Also fixed issue where framebuffer size was not being set on SDL

+++ code/shared/float_conversions.hpp
@@ -26,9 +26,9 @@ union FP16
     u16 u;
     struct
     {
-        u32 Mantissa : 10;
-        u32 Exponent : 5;
-        u32 Sign : 1;
+        u16 Mantissa : 10;
+        u16 Exponent : 5;
+        u16 Sign : 1;
     };
 };
